Make it buildable on modern GHC
The code uses FiniteMap (as it was back in 2004), but it is no
longer included in GHC. It's on Hackage, but `cabal install
FiniteMap` barfs. So I just grabbed FiniteMap.lhs from there,
un-literated it, updated the module name, and put it in this
repo.

Otherwise, it's mostly updating old 'flat' module names
(Monad -> Control.Monad), some changes in Curses.hsc to appease
clang, and some Makefile changes.